Rename: smoke -> e2e
